+++ src/constants/export/index.ts
@@ -1,1 +1,28 @@
 export const METADATA_FOLDER_NAME = 'metadata';
+
+export enum ExportNotification {
+    START = 'export started',
+    IN_PROGRESS = 'export already in progress',
+    FINISH = 'export finished',
+    FAILED = 'export failed',
+    ABORT = 'export aborted',
+    PAUSE = 'export paused',
+    UP_TO_DATE = `no new files to export`,
+}
+
+export enum RecordType {
+    SUCCESS = 'success',
+    FAILED = 'failed',
+}
+export enum ExportStage {
+    INIT,
+    INPROGRESS,
+    PAUSED,
+    FINISHED,
+}
+
+export enum ExportType {
+    NEW,
+    PENDING,
+    RETRY_FAILED,
+}
