+++ src/types/export/index.ts
@@ -1,17 +1,4 @@
-export enum ExportNotification {
-    START = 'export started',
-    IN_PROGRESS = 'export already in progress',
-    FINISH = 'export finished',
-    FAILED = 'export failed',
-    ABORT = 'export aborted',
-    PAUSE = 'export paused',
-    UP_TO_DATE = `no new files to export`,
-}
-
-export enum RecordType {
-    SUCCESS = 'success',
-    FAILED = 'failed',
-}
+import { ExportStage } from 'constants/export';
 
 export type CollectionIDPathMap = Map<number, string>;
 export interface ExportProgress {
@@ -35,16 +22,4 @@ export interface ExportRecord {
     exportedFiles?: string[];
     failedFiles?: string[];
     exportedCollectionPaths?: ExportedCollectionPaths;
-}
-export enum ExportStage {
-    INIT,
-    INPROGRESS,
-    PAUSED,
-    FINISHED,
-}
-
-export enum ExportType {
-    NEW,
-    PENDING,
-    RETRY_FAILED,
 }
